[refs #00005] Add LICENSE
This code needs licensing. It’s a bit of a sensitive/grey area right now
as the code was written by me before engaging with Sky, but was refined
in its current for whilst at Sky. As such, I’ve joint licensed it under
Apache 2.0. Speak to me if this needs changing.